complete setting up tailwind and react
