+++ src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-// import { BrowserRouter as Router, Ruotes, Route } from "react-router-dom";
+// import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
 import { Navbar } from "./components/navbar";
